# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Oil \u0026 Gas Service Booking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/services": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "services"
                ],
                "summary": "Получить все услуги",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/oil-gas-service-booking_internal_models.Service"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Создать услугу",
                "parameters": [
                    {
                        "description": "Данные услуги",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handlers.ServiceCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/oil-gas-service-booking_internal_models.Service"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/services/my": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Получить услуги моих компаний",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/oil-gas-service-booking_internal_models.Service"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/services/{id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "services"
                ],
                "summary": "Получить услугу по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oil-gas-service-booking_internal_models.Service"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "services"
                ],
                "summary": "Обновить услугу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oil-gas-service-booking_internal_models.Service"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "services"
                ],
                "summary": "Удалить услугу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_http-server_handlers.ServiceCreateRequest": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string",
                    "example": "Диагностика оборудования"
                }
            }
        },
...
        "oil-gas-service-booking_internal_models.Booking": {
            "type": "object",
            "properties": {
                "bookingID": {
                    "type": "integer"
                },
                "bookingServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.BookingService"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "oil-gas-service-booking_internal_models.BookingService": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.Booking"
                },
                "bookingID": {
                    "type": "integer"
                },
                "bookingServiceID": {
                    "type": "integer"
                },
                "companyService": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                },
                "companyServiceID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "oil-gas-service-booking_internal_models.Company": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "companyID": {
                    "type": "integer"
                },
                "companyServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.User"
                },
                "userID": {
                    "description": "владелец",
                    "type": "integer"
                }
            }
        },
        "oil-gas-service-booking_internal_models.CompanyService": {
            "type": "object",
            "properties": {
                "bookingServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.BookingService"
                    }
                },
                "company": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.Company"
                },
                "companyID": {
                    "type": "integer"
                },
                "companyServiceID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "price": {
                    "description": "цена конкретной услуги у компании",
                    "type": "number"
                },
                "service": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.Service"
                },
                "serviceID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "oil-gas-service-booking_internal_models.Service": {
            "type": "object",
            "properties": {
                "companyServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "serviceID": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "oil-gas-service-booking_internal_models.User": {
            "type": "object",
            "properties": {
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.Booking"
                    }
                },
                "companies": {
                    "description": "Добавляем обратную связь",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.Company"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}